Added Installation and Software section
Description:
Included additional sentences for the user to need a Code Editor and Python installed.  Bikeshare.py file should be located in same folder as this README.md file.
User instructed to open an active terminal within their Code Editor and type python bikeshare.py in the command line.
User shoud receive a welcome message from Bikeshare and also a prompt for user input for the city of choice for statistical data.