Added new function display_data
Description:
The display_data function requests input from user if they want raw data to be displayed.  If user response is 'yes', five rows of raw data is displayed with a variable that can be configured in future releases.
If user response is anything other than 'yes', a thank you message is displayed and the system exits.

After displaying the first five rows of raw data, the user is prompted to respond if they want more raw data displayed.  If response is 'yes', an additional five rows of raw data is displayed.  Any response other
than 'yes' results in a thank you message and the system exits.

Future releases may include the ability for the user to be prompted for if they want to do another search of the data with new filter criteria.