Fixed bug: script not providing data for Sunday
Description:
Additional testing identified data for Sunday was not populating due to +1 added to the weekday variable.  Removed the +1 and tested calculations and raw data output for user filters for Sunday.

Minor adjustment: Made additional spacing adjustments with '\n' to printed lines for informing user that the system is gathering information for their specific filter criteria.