refactor: Restart get_filters function for city
This refactor provides a prompt to the user if the city input does not match the available city data available instead of exiting the system.
Example: User is prompted to enter a city name for Chicago, New York City, or Washington.  If the user input is 'Boston', the user receives a message stating the city is not recognized. Do you want to restart?
If the user input is 'yes', the get_filters function is restarted and user is allowed to make another city input instead of having to restart the python script from the command prompt.

This refactor provides increased user experience of python script if the city is misspelled, wrong city is entered, or anyting is inputted other than the appropriate city name.